Fix use of platform specific defines, and remove unnecessary warnings.
Note that PATH_MAX is basically a made up number, as is the * 4 here.
MAX_PATH on windows is outdated (and tiny).

Effectively none of the macros can tell you the kernel or driver level limitations anyway, so this is a best effort guess based on the best effort guess that is prevalent already in c headers.

This particular constant usage can be looked at closer another time, but this is to fix a broken build.