Use submodules for dependencies: wren and libuv
This changes nothing about the projects themselves,
but saves disk space and reduces unessicarry changes in commits.

In actuality, the 'wren/include' directory moved to 'wren/src/include'
so we had to update all the build projects.
I think I'm running with a newer version of premake so the generated
makefiles changed very slightly too (more than would be expected).